Split Plane class out from Guide to general geometry.
